Create Python_Painting_Pikachu
Draw Pikachu with Python Turtle
First Part ( Import, getPositionl() and part 1 of the class ):
First, import the turtle module and create a function named “getPosition” with the parameters (x, y). Inside this function, set the x of the turtle to “x” and y of the turtle to “y.” Then, print the x and y.
Create a class named Pikachu. Inside this class, create an initializer __init__. Inside this function, set “t” as the turtle and set the pen size to 3 and speed to 9. it is not necessary that functions inside a class always have a parameter “self” I've given as ash you can write any, Now, on drag, call the getposition function.
Create another function with any name, I've named it Trace() with the parameters self, x, and y. inside this function, pick the pen up and go to (x, y), and put the pen down.
Likewise, create another function with any name I've named it leftEye() with the parameters ash, x, and y. Inside this function, call the Trace() function with the arguments (x, y). Set “ash.t” as “t”. Call the seth() method with the angle of 0 degrees and fill the part with the color#333333 and begin the fill. Draw a circle(22) and end the fill. Again, call the Trace() function with the arguments (x, y+10) and fill the part with the color #000000 ( black ), and begin the fill. Create a circle(10) and end the fill. Again, call the Trace() function with the arguments (x+6, y+22) and fill the part with the color #ffffff ( white ) and begin the fill. Create a circle(10) and end the fill.
Similarly, create another function named rightEye() with the parameters ash, x, and y. Inside this function, call the Trace() function with the arguments x and y. Then, set “ash.t” as “t.” Now, call the seth() method with the angle of 0 and fill the part with the color #333333 and begin the fill. Draw a circle(22) and end the fill. Continuing, call the Trace() function with the arguments x, y+19 and fill the part with the color #00000 ( black ) and begin the fill. Draw a circle(10) and end the fill. Similarly, call the Trace() function again with the arguments (x-6, y + 22) and fill the part with the color #ffffff ( white ) and begin the fill. Draw a circle(10) and end the flll.

Second Part ( Class Part – 2 ):
Inside the same class, create a function name mouth() with the parameters self, x, y. Inside this function, call the Trace(x, y) and fill the part with the color #88141D and begin the fill. Create two empty lists with the variables named l1 and l2. Call the seth(190) and (a = 0.7). Create a for loop with the range of 28 and add 0.1 to a and move the turtle right(3), fd(a), and append to l1 the turtle’s position. Call the Turtle(x, y) and seth(10) and (a =0.7). Create another for loop with the range of 28 and do the same as above but with l2. Then call the seth(10), circle(50, 15), Left(180) and circle(-50, 15) for as in the code above in the line 96. After calling the Trace(17, 54) again and fill the part with the color#DD716F and begin the fill. Call the seth() method with the 145 angles, circle(40, 86), and pick the pen up. Now, create a for loop at the range of {reversed(l1[:20])}; goto { t.goto(pos[0], pos[1]+1.5) }. Again create a for loop with the range of { l2[:20] }; and go to { t.goto(pos[0], pos[1]+1.5) }.
Again, call a function named leftCheek with the parameters of self, x, y. Inside this function, set the tracer to false and call the Trace(x, y), seth(300), fill the part with the color #dd4d28 begin the fill. Set the value of “a” to 2.3 and create a for loop with the range of 120; create an if function which will sub 0.05 from a and lt(3),fd(a) of I is between 0 and 30 or 60 and 90. else; add 0.05 to a and lt(3),fd(a).

Third Part ( Class Part – 3 ):
Moreover, create a function rightcheek(self, x, y). Inside this function, call the Trace(x, y), seth(60), fill the color #DD4D28 and begin the fill and do the same as above with rightcheecks() as in leftcheeks().
Then, create a colorLeftEar(self, x, y) function; call the Trace(x, y), fill the color black and begin the fill. call t.seth(330), t.circle(100, 35), t.seth(219), t.circle(-300, 19),t.seth(110), t.circle(-30, 50), t.circle(-300, 10), t.end_fill(). Do the same with colorRightEar()
Create a jiu() function; fill the color as #f6202f and begin the fill. Pick the pen up and circle(130, 40). Put the pen down and draw a circle(100, 105), move left at an angle of 180, and draw another circle(-100, 5). The values for the methods will change but do the code as above.

 
Last Function in the class:
Create a cap() function with the parameters of self,x,y. Inside this function, call Trace(x, y) and fill the color #CD0000 and begin the fill. Call seth(200),t.circle(400, 7),t.left(180), t.circle(-400, 30),t.circle(30, 60),t.fd(50),t.circle(30, 45), t.fd(60),t.left(5), t.circle(30, 70), t.right(20),t.circle(200, 70),t.circle(30, 60), t.fd(70), t.right(35),t.fd(50),t.circle(8, 100),t.end_fill(),self.meme(-168.47, 185.52),t.seth(36) ,t.circle(-270, 54),t.left(180),t.circle(270, 27),t.circle(-80, 98).
Create a function name start and call the function start() inside it.

Last Part:
Lastly, create a function main. Inside this function, set the screen size to 800 / 600. Then, create an object cartoon with the Cartoon() class and call the function start() through the cartoon object. Close the mainloop as in Tkinter.
At last, create a module source as this is the main program and call the main() function inside it.